refactor: extract chart trend logic and split updateButton into focused methods
- Extract ~150 lines of trend computation from ActivityChartView (678→526 lines)
  into ActivityChartTrend.swift: ActivityTrendComputation enum with pure comparison
  logic (vs yesterday, vs last week, vs last month) and shared formatting helpers
- Split StatusBarManager.updateButton (81 lines) into 6 focused methods:
  resolveMetricMode, resolveStarColor, updateSparkleState, updateRenderState,
  updateBreathTimer, resolveDisplayText — each under 10 lines
- Make compactCount and formatHourLabel internal for cross-file access
